Rename vendors to vendor_management
Since we changed the URL for the plugin it seemed like it made sense to
also change the module name. Also since most of the URL paths have been
pluralized (vendors, interfaces) I thought it made sense to pluralize
file as well for consistency.
